Extract contract test logic into base class
We are adding a second set of contract tests, so pulling all the code that can
be shared from the existing `TeacherTrainingAdviserTests` into a `BaseTests`
class.
